+++ src/backend/ci/core/process/biz-engine/src/main/kotlin/com/tencent/devops/process/engine/control/BuildEndControl.kt
@@ -53,7 +53,6 @@ import com.tencent.devops.process.engine.control.lock.PipelineBuildNoLock
 import com.tencent.devops.process.engine.control.lock.PipelineBuildStartLock
 import com.tencent.devops.process.engine.pojo.BuildInfo
 import com.tencent.devops.process.engine.pojo.LatestRunningBuild
-import com.tencent.devops.process.engine.pojo.PipelineBuildTask
 import com.tencent.devops.process.engine.pojo.event.PipelineBuildAtomTaskEvent
 import com.tencent.devops.process.engine.pojo.event.PipelineBuildFinishEvent
 import com.tencent.devops.process.engine.pojo.event.PipelineBuildStartEvent
@@ -252,10 +251,6 @@ class BuildEndControl @Autowired constructor(
                         )
                     )
                 }
-                // 如果是取消的构建，则会统一取消子流水线的构建
-                if (buildStatus.isPassiveStop() || buildStatus.isCancel()) {
-                    terminateSubPipeline(buildInfo.buildId, it)
-                }
             }
             // 将插件出错信息逐一加入构建错误信息
             if (it.errorType != null) {
@@ -275,52 +270,6 @@ class BuildEndControl @Autowired constructor(
             }
         }
         if (errorInfos.isNotEmpty()) buildInfo.errorInfoList = errorInfos
-    }
-
-    private fun terminateSubPipeline(buildId: String, buildTask: PipelineBuildTask) {
-
-        if (buildTask.subBuildId.isNullOrBlank()) {
-            return
-        }
-
-        val subBuildInfo = pipelineRuntimeService.getBuildInfo(buildTask.subBuildId!!)
-
-        if (subBuildInfo?.status?.isFinish() == false) { // 子流水线状态为未构建结束的，开始下发退出命令
-            try {
-                val tasks = pipelineRuntimeService.getRunningTask(subBuildInfo.projectId, subBuildInfo.buildId)
-                tasks.forEach { task ->
-                    val taskId = task["taskId"] ?: ""
-                    val containerId = task["containerId"] ?: ""
-                    val executeCount = task["executeCount"] ?: 1
-                    buildLogPrinter.addYellowLine(
-                        buildId = buildId,
-                        message = "Cancelled by pipeline[${buildTask.pipelineId}]，Operator:${buildTask.starter}",
-                        tag = taskId.toString(),
-                        jobId = containerId.toString(),
-                        executeCount = executeCount as Int
-                    )
-                }
-
-                if (tasks.isEmpty()) {
-                    buildLogPrinter.addYellowLine(
-                        buildId = buildId,
-                        message = "cancelled by pipeline[${buildTask.pipelineId}]，Operator:${buildTask.starter}",
-                        tag = "",
-                        jobId = "",
-                        executeCount = 1
-                    )
-                }
-                pipelineRuntimeService.cancelBuild(
-                    projectId = subBuildInfo.projectId,
-                    pipelineId = subBuildInfo.pipelineId,
-                    buildId = subBuildInfo.buildId,
-                    userId = subBuildInfo.startUser,
-                    buildStatus = BuildStatus.CANCELED
-                )
-            } catch (ignored: Exception) {
-                LOG.warn("ENGINE|$buildId|TerminateSubPipeline|subBuildId=${subBuildInfo.buildId}|e=$ignored")
-            }
-        }
     }
 
     private fun PipelineBuildFinishEvent.popNextBuild() {
